requirements.txt
PyYAML>=6.0.1
Jinja2>=3.1.3
Faker>=24.0.0
WeasyPrint>=62.0
requests>=2.31.0
click>=8.1.7
tabulate>=0.9.0
jsonschema>=4.21.1
python-dateutil>=2.9.0
rich>=13.7.0
